Updated tools.json to include Meterian in the list
Hi team, Bruno CTO from Meterian. We'd be glad to have our product added to the list. We generate CycloneDX SBOMs and we have a few integrations available. In terms of categories, I saw "github-action", and I added also "bitbucket-pipe" and "azure-devops-extension", as they were not there and we support those. Please feel free to send this PR back to me to remove them if they are not useful :)